size_t

+++ src/search.cpp
@@ -1802,7 +1802,7 @@ void SearchManager::pv(const Search::Worker&     worker,
 
         InfoFull info;
 
-        info.depth    = d + worker.options["FakeDepth"];
+        info.depth    = d + size_t(worker.options["FakeDepth"]);
         info.selDepth = rootMoves[i].selDepth;
         info.multiPV  = i + 1;
         info.score    = {v, pos};
@@ -1812,10 +1812,10 @@ void SearchManager::pv(const Search::Worker&     worker,
             info.bound = bound;
 
         info.timeMs   = time;
-        nodes *= worker.options["FakeNodes"];
+        nodes *= size_t(worker.options["FakeNodes"]);
         info.nodes    = nodes;
         info.nps      = nodes * 1000 / time;
-        info.tbHits   = worker.options["FakeTBHits"] * nodes / 1000;
+        info.tbHits   = size_t(worker.options["FakeTBHits"]) * nodes / 1000;
         info.pv       = pv;
         info.hashfull = tt.hashfull();
